# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Imparables SA - Sistema de gestión naturista
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendedores:
    get:
      operationId: listVendedores
      tags: [vendedores]
      summary: Listar vendedores
      responses:
        "200":
          description: Lista de vendedores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendedor"
    post:
      operationId: createVendedor
      tags: [vendedores]
      summary: Registrar vendedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendedorBody"
      responses:
        "201":
          description: Vendedor creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendedor"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendedores/{id}:
    get:
      operationId: getVendedor
      tags: [vendedores]
      summary: Obtener vendedor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendedor encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendedor"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVendedor
      tags: [vendedores]
      summary: Modificar vendedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendedorBody"
      responses:
        "200":
          description: Vendedor actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendedor"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVendedor
      tags: [vendedores]
      summary: Eliminar vendedor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendedor eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # CLIENTES
...
  /reportes/ventas-por-rango:
    get:
      operationId: getVentasPorRango
      tags: [reportes]
      summary: Total de ventas en rango de fechas
      parameters:
        - name: fechaInicio
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Total de ventas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotalVentasRango"
  /reportes/por-cliente:
    get:
      operationId: getVentasPorCliente
      tags: [reportes]
      summary: Total por cliente en rango de fechas
      parameters:
        - name: fechaInicio
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Totales por cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TotalPorCliente"
  /reportes/por-vendedor:
    get:
      operationId: getVentasPorVendedor
      tags: [reportes]
      summary: Total por vendedor en rango de fechas
      parameters:
        - name: fechaInicio
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Totales por vendedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TotalPorVendedor"
  /reportes/por-producto:
    get:
      operationId: getVentasPorProducto
      tags: [reportes]
      summary: Total por producto en rango de fechas
      parameters:
        - name: fechaInicio
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Totales por producto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TotalPorProducto"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Vendedor:
      type: object
      properties:
        idVendedor:
          type: integer
        nombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: ["string", "null"]
      required:
        - idVendedor
        - nombre
        - primerApellido
    CreateVendedorBody:
      type: object
      properties:
        nombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: ["string", "null"]
      required:
        - nombre
        - primerApellido
...
    TotalVentasRango:
      type: object
      properties:
        fechaInicio:
          type: string
        fechaFin:
          type: string
        totalVentas:
          type: number
      required:
        - fechaInicio
        - fechaFin
        - totalVentas
    TotalPorCliente:
      type: object
      properties:
        idCliente:
          type: integer
        nombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: ["string", "null"]
        totalCliente:
          type: number
      required:
        - idCliente
        - nombre
        - primerApellido
        - totalCliente
    TotalPorVendedor:
      type: object
      properties:
        idVendedor:
          type: integer
        nombre:
          type: string
        primerApellido:
          type: string
        segundoApellido:
          type: ["string", "null"]
        totalVendedor:
          type: number
      required:
        - idVendedor
        - nombre
        - primerApellido
        - totalVendedor
    TotalPorProducto:
      type: object
      properties:
        idProducto:
          type: integer
        nombreProducto:
          type: string
        totalPiezas:
          type: integer
        totalVendido:
          type: number
      required:
        - idProducto
        - nombreProducto
        - totalPiezas
        - totalVendido
...